* commit '9f91e23e5646d65c899ddb415ec6bf540d43cc0f':
  Login implemented with the common login solution
  Fixed problems
  Rewrote login handling to use the custom login

+++ swords-user/src/main/java/com/morethanheroic/swords/user/repository/domain/UserMapper.java
@@ -21,22 +21,22 @@ public interface UserMapper {
     UserDatabaseEntity findByUsernameAndPassword(String username, String password);
 
     @Select("SELECT * FROM users WHERE id = #{id}")
-    UserDatabaseEntity findById(int id);
+    UserDatabaseEntity findById(long id);
 
     @Update("UPDATE users SET last_login_date = NOW() WHERE id = #{id}")
-    void updateLastLoginDate(@Param("id") int id);
+    void updateLastLoginDate(@Param("id") long id);
 
     @Update("UPDATE users SET scavenging_point= #{scavenging_point} WHERE id = #{userId}")
-    void updateScavengingPoint(@Param("userId") int userId, @Param("scavenging_point") int scavengingPoint);
+    void updateScavengingPoint(@Param("userId") long userId, @Param("scavenging_point") int scavengingPoint);
 
     @Update("UPDATE users SET movement = #{movement} WHERE id = #{userId}")
-    void updateMovement(@Param("userId") int userId, @Param("movement") int movement);
+    void updateMovement(@Param("userId") long userId, @Param("movement") int movement);
 
     @Update("UPDATE users SET health = #{health}, mana = #{mana}, movement = #{movement} WHERE id = #{userId}")
-    void updateBasicCombatStats(@Param("userId") int userId, @Param("health") int health, @Param("mana") int mana, @Param("movement") int movement);
+    void updateBasicCombatStats(@Param("userId") long userId, @Param("health") int health, @Param("mana") int mana, @Param("movement") int movement);
 
     @Update("UPDATE users SET last_regeneration_date = #{date}, health = #{health}, mana = #{mana}, movement = #{movement} WHERE id = #{userId}")
-    void updateRegeneration(@Param("userId") int userId, @Param("health") int health, @Param("mana") int mana,
+    void updateRegeneration(@Param("userId") long userId, @Param("health") int health, @Param("mana") int mana,
                             @Param("movement") int movement, @Param("date") Instant date);
 
     @Options(useGeneratedKeys = true, keyProperty = "id", keyColumn = "id")
